Now generate the image also.
